# lib/protocol: Optimize luhn and chunk functions