@st-review 2017-09-03
:ok_hand: Merged as 9682bbfbdacf44e86d788da2065a0094dc70cec9. Thanks, @calmh!